+++ lib/packages/wikibase/data-model/RELEASE-NOTES.md
@@ -41,6 +41,7 @@ Other breaking changes:
 * `Item::removeSiteLink` no longer takes an optional second parameter and no longer returns a boolean
 * Removed `Fingerprint::getAliases`
 * Removed `EntityId::newFromPrefixedId`
+* The constructor of `EntityId` is no longer public
 
 #### Additions
 
@@ -51,6 +52,7 @@ Other breaking changes:
 * Added `SiteLinkList::removeLinkWithSiteId`
 * Added `Item::getSiteLinkList`
 * Added `Item::setSiteLinkList`
+* Added `LegacyIdInterpreter`
 
 ## Version 0.7.4 (2014-04-24)